More project structure fix-up
