+++ .gitignore
@@ -69,7 +69,6 @@ fabric.properties
 [Ll]ocal
 [Ss]cripts
 pyvenv.cfg
-.venv
 pip-selfcheck.json
 ### Python template
 # Byte-compiled / optimized / DLL files
