+++ .travis.yaml
@@ -1,6 +1,6 @@
 language: python
 python:
-  - "3.5"
+  - "3.7"
 install:
   # Install test dependencies
   - pip install .[test]
